+++ Cunknown/test.c
@@ -10,7 +10,7 @@ int testFonctions()
     int rot[nboutput];
     pcg128_t S0 = (((pcg128_t) 5995207026785010249u) << k) + ((pcg128_t) 179350442155841024u);
     pcg128_t vraiS[nboutput];
-    unsigned long long X[nboutput];
+    u64 X[nboutput];
     pcg128_t c = ((((pcg128_t) 6364136223846793005u) << k) + 1442695040888963407u) >> 1;
     pcg(vraiS, X, S0, &c, nboutput);
     if(vraiS[2] != (((pcg128_t) 1792771836637573954u) << k) + ((pcg128_t) 11139816115278170276u)){
@@ -21,15 +21,15 @@ int testFonctions()
         rot[i] = (int) (vraiS[i] >> (2 * k - known_up));
     }
     
-    unsigned long long W0 = (unsigned long long) (vraiS[0] % (1<<known_low));
-    unsigned long long WC = (unsigned long long) (c % (1<<known_low));
+    u64 W0 = (u64) (vraiS[0] % (1<<known_low));
+    u64 WC = (u64) (c % (1<<known_low));
  
     printf("1..7\n");
     printf("# nbiter = %d\n", nbiter);
     printf("# known_low = %d\n", known_low);
 
     /* test unrotate */
-    unsigned long long uX[nbiter];
+    u64 uX[nbiter];
     unrotateX(uX, X, rot);
     if (uX[0] != 15007519919903780682u) {
         printf("not ok 1 - erreur sur unrotateX\n");
@@ -38,7 +38,7 @@ int testFonctions()
     }
 
     /* test getY */
-    unsigned long long Y[nbiter];
+    u64 Y[nbiter];
     printf("# W0 = %llu\n", W0);
     printf("# Wc = %llu\n", WC);
 
@@ -50,7 +50,7 @@ int testFonctions()
     }
 
     /* test getYprim */
-    unsigned long long Yprim[nbiter];
+    u64 Yprim[nbiter];
     getYprim(Yprim, Y, W0, WC);
     if(Yprim[1] != 93486){
         printf("not ok 3 - erreur sur getY\n");
@@ -59,7 +59,7 @@ int testFonctions()
     }
 
     /* test getDY */
-    unsigned long long DY[nbiter-1];
+    u64 DY[nbiter-1];
     getDY(DY, Yprim);
     if(DY[0] != 14609){
         printf("not ok 4 - erreur sur getDY\n");
@@ -68,21 +68,21 @@ int testFonctions()
     }
 
     /**** Polynômes en WC et W0 utilisés dans la résolution ****/
-    unsigned long long lowSumPol[nbiter + nbtest];
-    unsigned long long sumPolY[nbiter];
-    unsigned long long sumPolTest[nbtest];
+    u64 lowSumPol[nbiter + nbtest];
+    u64 sumPolY[nbiter];
+    u64 sumPolTest[nbtest];
     for(int i = 0 ; i < nbiter ; i++){
-        lowSumPol[i]  = (W0 * ((unsigned long long) powA[i]) + WC * ((unsigned long long) polA[i]));
+        lowSumPol[i]  = (W0 * ((u64) powA[i]) + WC * ((u64) polA[i]));
         sumPolY[i] = (polA[i] * WC + powA[i] * W0) >> (k - known_up);
     }
     for(int i = 0 ; i < nbtest ; i++){
-        lowSumPol[nbiter + i] = (W0 * ((unsigned long long) powA[i + nbiter]) + WC * ((unsigned long long) polA[i + nbiter]));
-        sumPolTest[i] = W0 * ((unsigned long long) (powA[i + nbiter] >> known_low) - 1) + WC * ((unsigned long long) (polA[i + nbiter] >> known_low) - 1);
+        lowSumPol[nbiter + i] = (W0 * ((u64) powA[i + nbiter]) + WC * ((u64) polA[i + nbiter]));
+        sumPolTest[i] = W0 * ((u64) (powA[i + nbiter] >> known_low) - 1) + WC * ((u64) (polA[i + nbiter] >> known_low) - 1);
     }
     
 
     /* test FindDS64 */
-    unsigned long long DS64[nbiter - 1];
+    u64 DS64[nbiter - 1];
     FindDS64(DS64, Y, uX, rot, lowSumPol, sumPolY);
     if(DS64[0] != 2055999906439120392u){
         printf("not ok 5 - erreur sur FindDS64\n");
@@ -110,18 +110,18 @@ int testFonctions()
     char* goodY = setupGoodY();
     getGoodY(goodY, X, lowSumPol, 1);
 
-    unsigned long long tabTmp[k * nbiter];
+    u64 tabTmp[k * nbiter];
     getTabTmp(tabTmp, X, lowSumPol, sumPolY);
 
-    // unsigned long long uXnbiter1 = unrotate(X[nbiter + 1], rot[nbiter + 1]);
-    // unsigned long long Ynbiter1 = ((((unsigned long long) ((polA[nbiter + 1] * WC + powA[nbiter + 1] * W0) % (1 << known_low))) ^ (uXnbiter1 % (1 << known_low))) << known_up) + (rot[nbiter + 1] ^ (uXnbiter1 >> (k - known_up)));
+    // u64 uXnbiter1 = unrotate(X[nbiter + 1], rot[nbiter + 1]);
+    // u64 Ynbiter1 = ((((u64) ((polA[nbiter + 1] * WC + powA[nbiter + 1] * W0) % (1 << known_low))) ^ (uXnbiter1 % (1 << known_low))) << known_up) + (rot[nbiter + 1] ^ (uXnbiter1 >> (k - known_up)));
     // if (!checkY(goodY, 1, Ynbiter1)) {
     // //if(!goodY[Ynbiter1 + (1<<(known_low + known_up))]){
     //     printf("not ok 6 - erreur sur getGoodY\n");
     // }  else {
     //     printf("ok 6 - getGoodY\n");
     // }
-    unsigned long long DS640, Y0;
+    u64 DS640, Y0;
     if(!solve_isgood(goodY, rot, tabTmp, sumPolY, sumPolTest)){
         printf("not ok 6 - erreur sur solve_isgood\n");
     }  else {
@@ -139,23 +139,23 @@ int testFonctions()
 void printVal(pcg128_t S0, pcg128_t c){
     int rot[nboutput];
     pcg128_t vraiS[nboutput];
-    unsigned long long X[nboutput];
+    u64 X[nboutput];
     pcg(vraiS, X, S0, &c, nboutput);
-    printf("setseq : %llu %llu\n", (unsigned long long) (vraiS[2]>>64), (unsigned long long) vraiS[2]);
+    printf("setseq : %llu %llu\n", (u64) (vraiS[2]>>64), (u64) vraiS[2]);
     
-    //printf("%llu %llu\n", (unsigned long long) ((vraiS[1] - vraiS[0])>>64), (unsigned long long) (vraiS[1] - vraiS[0]));
+    //printf("%llu %llu\n", (u64) ((vraiS[1] - vraiS[0])>>64), (u64) (vraiS[1] - vraiS[0]));
     
     //done
     for(int i = 0 ; i < nboutput ; i++){
         rot[i] = (int) (vraiS[i] >> (2 * k - known_up));
     }
     
-    unsigned long long W0 = (unsigned long long) (vraiS[0] % (1<<known_low));
-    unsigned long long WC = (unsigned long long) (c % (1<<known_low));
+    u64 W0 = (u64) (vraiS[0] % (1<<known_low));
+    u64 WC = (u64) (c % (1<<known_low));
     printf("W0 : %llu\n", W0);
     printf("WC : %llu\n", WC);
     
-    unsigned long long uX[nbiter];
+    u64 uX[nbiter];
     unrotateX(uX, X, rot);
     printf("uX\n");
     for(int i = 0 ; i < nbiter ; i++)
@@ -165,7 +165,7 @@ void printVal(pcg128_t S0, pcg128_t c){
     f = fopen("result.txt","w");
     
     fprintf(f,"W0 : %llu\n", W0);*/
-    unsigned long long Y[nbiter];
+    u64 Y[nbiter];
     getY(Y, W0, WC, rot, uX);
     printf("Y :\n");
     for(int i = 0 ; i < nbiter ; i++)
@@ -174,14 +174,14 @@ void printVal(pcg128_t S0, pcg128_t c){
     
     
     /***** Tests de vérification des sous-fonctions *****/
-    unsigned long long Yprim[nbiter];
+    u64 Yprim[nbiter];
     getYprim(Yprim, Y, W0, WC);
     printf("Yprim :\n");
     for(int i = 0 ; i < nbiter ; i++)
         printf("%llu ", Yprim[i]);
     printf("\n");
     
-    unsigned long long DY[nbiter];
+    u64 DY[nbiter];
     getDY(DY, Yprim);
     printf("DY :\n");
     for(int i = 0 ; i < nbiter - 1 ; i++)
@@ -189,19 +189,19 @@ void printVal(pcg128_t S0, pcg128_t c){
     printf("\n");
 
     /**** Polynômes en WC et W0 utilisés dans la résolution ****/
-    unsigned long long lowSumPol[nbiter + nbtest];
-    unsigned long long sumPolY[nbiter];
-    unsigned long long sumPolTest[nbtest];
+    u64 lowSumPol[nbiter + nbtest];
+    u64 sumPolY[nbiter];
+    u64 sumPolTest[nbtest];
     for(int i = 0 ; i < nbiter ; i++){
-        lowSumPol[i]  = (W0 * ((unsigned long long) powA[i]) + WC * ((unsigned long long) polA[i]));
+        lowSumPol[i]  = (W0 * ((u64) powA[i]) + WC * ((u64) polA[i]));
         sumPolY[i] = (polA[i] * WC + powA[i] * W0) >> (k - known_up);
     }
     for(int i = 0 ; i < nbtest ; i++){
-        lowSumPol[nbiter + i] = (W0 * ((unsigned long long) powA[i + nbiter]) + WC * ((unsigned long long) polA[i + nbiter]));
-        sumPolTest[i] = W0 * ((unsigned long long) (powA[i + nbiter] >> known_low) - 1) + WC * ((unsigned long long) (polA[i + nbiter] >> known_low) - 1);
+        lowSumPol[nbiter + i] = (W0 * ((u64) powA[i + nbiter]) + WC * ((u64) polA[i + nbiter]));
+        sumPolTest[i] = W0 * ((u64) (powA[i + nbiter] >> known_low) - 1) + WC * ((u64) (polA[i + nbiter] >> known_low) - 1);
     }
     
-    unsigned long long DS64[nbiter - 1];
+    u64 DS64[nbiter - 1];
     FindDS64(DS64, Y, uX, rot, lowSumPol,sumPolY);
     printf("DS64 :\n");
     for(int i = 0 ; i < nbiter - 1 ; i++)
