+++ Cunknown/test_falsenegative.c
@@ -7,7 +7,7 @@ int testValid (FILE* f, int n)
 {
     int rot[nboutput];
     pcg128_t vraiS[nboutput];
-    unsigned long long X[nboutput];
+    u64 X[nboutput];
     pcg128_t seeds[2];
     int cpt = 0;
     char* goodY = setupGoodY();
@@ -23,33 +23,33 @@ int testValid (FILE* f, int n)
         for(int i = 0 ; i < nboutput ; i++)
             rot[i] = (int) (vraiS[i] >> (2 * k - known_up));
 
-        unsigned long long W0 = (unsigned long long) (vraiS[0] % (1<<known_low));
-        unsigned long long WC = (unsigned long long) (seeds[1] % (1<<known_low));
+        u64 W0 = (u64) (vraiS[0] % (1<<known_low));
+        u64 WC = (u64) (seeds[1] % (1<<known_low));
         
         /**** Polynômes en WC et W0 utilisés dans la résolution ****/
-        unsigned long long lowSumPol[nbiter + nbtest];
-        unsigned long long sumPolY[nbiter];
-        unsigned long long sumPolTest[nbtest];
+        u64 lowSumPol[nbiter + nbtest];
+        u64 sumPolY[nbiter];
+        u64 sumPolTest[nbtest];
         for(int i = 0 ; i < nbiter ; i++){
-            lowSumPol[i]  = (W0 * ((unsigned long long) powA[i]) + WC * ((unsigned long long) polA[i]));
+            lowSumPol[i]  = (W0 * ((u64) powA[i]) + WC * ((u64) polA[i]));
             sumPolY[i] = (polA[i] * WC + powA[i] * W0) >> (k - known_up);
         }
         for(int i = 0 ; i < nbtest ; i++){
-            lowSumPol[nbiter + i] = (W0 * ((unsigned long long) powA[i + nbiter]) + WC * ((unsigned long long) polA[i + nbiter]));
-            sumPolTest[i] = W0 * ((unsigned long long) (powA[i + nbiter] >> known_low) - 1) + WC * ((unsigned long long) (polA[i + nbiter] >> known_low) - 1);
+            lowSumPol[nbiter + i] = (W0 * ((u64) powA[i + nbiter]) + WC * ((u64) polA[i + nbiter]));
+            sumPolTest[i] = W0 * ((u64) (powA[i + nbiter] >> known_low) - 1) + WC * ((u64) (polA[i + nbiter] >> known_low) - 1);
         }
 
-        unsigned long long DS640;
-        unsigned long long Y0;
-        /*unsigned long long uX[nbiter];
+        u64 DS640;
+        u64 Y0;
+        /*u64 uX[nbiter];
         unrotateX(uX, X, rot);
         
-        unsigned long long Y[nbiter]; //utilisé dans testDS640
+        u64 Y[nbiter]; //utilisé dans testDS640
         getY(Y, W0, WC, rot, uX);
         FindDS64(DS64, Y, uX, rot, lowSumPol, sumPolY);
         cpt += testDS640(DS64[0], X, Y[0], sumPolTest, lowSumPol);*/
 
-        unsigned long long tabTmp[k * nbiter];
+        u64 tabTmp[k * nbiter];
         getTabTmp(tabTmp, X, lowSumPol, sumPolY);    
 
         getGoodY(goodY, X, lowSumPol, 1);
